requirements.txt
xrpl-py>=2.0.0
websockets>=10.4
aiohttp>=3.8.6
python-dotenv>=1.0.0
typing-extensions>=4.5.0  # Required by xrpl-py
base58>=2.1.1  # Required by xrpl-py
ECPy>=1.2.5  # Required by xrpl-py
httpx>=0.23.3  # Required by xrpl-py
PyYAML>=6.0.1  # For configuration management

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # For code coverage reporting